Move WriteVTK code to an extension
In an effort to minimize Raven's dependencies, we move the WriteVTK code
to a Pkg extension.  This fixes #20.
