+++ src/grids.jl
@@ -8,33 +8,6 @@ floattype(grid::AbstractGrid) = floattype(typeof(grid))
 arraytype(grid::AbstractGrid) = arraytype(typeof(grid))
 celltype(grid::AbstractGrid) = celltype(typeof(grid))
 
-function WriteVTK.vtk_grid(filename::AbstractString, grid::AbstractGrid, args...; kwargs...)
-    return pvtk_grid(filename, grid, args...; kwargs...)
-end
-
-function WriteVTK.pvtk_grid(
-    filename::AbstractString,
-    grid::AbstractGrid,
-    args...;
-    withghostlayer = Val(false),
-    kwargs...,
-)
-    part = partitionnumber(grid)
-    nparts = numberofpartitions(grid)
-    vtk = pvtk_grid(
-        filename,
-        points_vtk(grid, withghostlayer),
-        cells_vtk(grid, withghostlayer),
-        args...;
-        part,
-        nparts,
-        kwargs...,
-    )
-    data_vtk!(vtk, grid, withghostlayer)
-
-    return vtk
-end
-
 struct Grid{C<:AbstractCell,P,L,T,F,PN,N,CTOD,DTOC} <: AbstractGrid{C}
     part::Int
     nparts::Int
@@ -85,38 +58,3 @@ numcells(grid::Grid, ::Val{true}) = length(grid.levels)
 Base.length(grid::Grid) = grid.locallength
 partitionnumber(grid::Grid) = grid.part
 numberofpartitions(grid::Grid) = grid.nparts
-
-function points_vtk(grid::Grid, withghostlayer = Val(false))
-    P = toequallyspaced(referencecell(grid))
-    x = P * reshape(points(grid, withghostlayer), size(P, 2), :)
-
-    return reinterpret(reshape, eltype(eltype(x)), vec(adapt(Array, x)))
-end
-
-function cells_vtk(grid::Grid, withghostlayer = Val(false))
-    type = celltype_vtk(referencecell(grid))
-    connectivity = connectivity_vtk(referencecell(grid))
-
-    cells = [
-        MeshCell(type, e * length(connectivity) .+ connectivity) for
-        e = 0:(numcells(grid, withghostlayer)-1)
-    ]
-
-    return cells
-end
-
-function data_vtk!(vtk, grid::Grid, withghostlayer = Val(false))
-    higherorderdegrees = zeros(Int, 3, numcells(grid, withghostlayer))
-    ds = [degrees(referencecell(grid))...]
-    higherorderdegrees[1:length(ds), :] .= repeat(ds, 1, numcells(grid, withghostlayer))
-
-    vtk["HigherOrderDegrees", VTKCellData()] = higherorderdegrees
-    vtk[VTKCellData()] = Dict("HigherOrderDegrees" => "HigherOrderDegrees")
-
-    vtk["PartitionNumber", VTKCellData()] =
-        fill(partitionnumber(grid), numcells(grid, withghostlayer))
-    vtk["Level", VTKCellData()] = collect(levels(grid, withghostlayer))
-    vtk["Tree", VTKCellData()] = collect(trees(grid, withghostlayer))
-
-    return
-end
