+++ test/Project.toml
@@ -2,5 +2,6 @@
 CUDA = "052768ef-5323-5732-b1bb-66c8b64840ba"
 MPI = "da04e1cc-30fd-572f-bb4f-1f8673147195"
 StableRNGs = "860ef19b-820b-49d6-a774-d7a799459cd3"
+WriteVTK = "64499a7a-5c06-52f2-abe2-ccb03c286192"
 Pkg = "44cfe95a-1eb2-52ea-b672-e2afdf69b78f"
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
